UPDATE LOGIC FOR NEXT HASH ATTRIBUTES
* Removed init and restart_init function
Removed functions as it is not required with the new logic
* Removed pruneheight and blocks variable
pruneheight and blocks variable is replaced with height. Height is
maintaining the height of blocks in the new logic and control the
blockchaininfo function.
* Updated dataProcessing function for the variable
DataProcessing function is updated to accept the hash variable. If
next block hash is passed, the function will get the block info
using hash else it will use the height to get the block hash and
get the block details after that.
* Timeout is set to 10 minutes
Timeout is set to 10 minutes as it takes 10 minutes on an average
to process a block in bitcoin.